+int to float dropchance

+++ Assets/Code/Scripts/Loot.cs
@@ -8,9 +8,9 @@ public class Loot : ScriptableObject
 {
     public Sprite lootSprite;
     public string lootName;
-    public int dropChance;
+    public float dropChance;
     
-    public Loot(string lootName, int dropChance)
+    public Loot(string lootName, float dropChance)
     {
         this.lootName = lootName;
         this.dropChance = dropChance;
